refactor: payment 도메인 위치 수정 - _payment -> payment #177
