--- src/main/java/poomasi/domain/order/_payment/service/FarmPaymentService.java
@@ -1,15 +1,0 @@
-package poomasi.domain.order._payment.service;
-
-
-import com.siot.IamportRestClient.exception.IamportResponseException;
-import org.springframework.stereotype.Service;
-
-import java.io.IOException;
-
-@Service
-public class FarmPaymentService {
-
-
-
-
-}
