+++ src/main/java/poomasi/domain/order/controller/OrderController.java
@@ -8,9 +8,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.web.bind.annotation.*;
-import poomasi.domain.farm.service.FarmService;
-import poomasi.domain.order._payment.dto.request.PaymentPreRegisterRequest;
-import poomasi.domain.order._payment.service.ProductPaymentService;
+import poomasi.payment.dto.request.PaymentPreRegisterRequest;
+import poomasi.payment.service.ProductPaymentService;
 import poomasi.domain.order.dto.request.ProductOrderRegisterRequest;
 import poomasi.domain.order.service.FarmOrderService;
 import poomasi.domain.order.service.ProductOrderService;
