+++ src/main/java/poomasi/payment/dto/request/PaymentPreRegisterRequest.java
@@ -1,4 +1,4 @@
-package poomasi.domain.order._payment.dto.request;
+package poomasi.payment.dto.request;
 
 import java.math.BigDecimal;
