+++ src/main/java/poomasi/payment/dto/request/PaymentValidateRequest.java
@@ -1,4 +1,4 @@
-package poomasi.domain.order._payment.dto.request;
+package poomasi.payment.dto.request;
 
 public record PaymentValidateRequest(String merchantUid, String amount) {
 }
