+++ src/main/java/poomasi/payment/entity/PaymentMethod.java
@@ -1,4 +1,4 @@
-package poomasi.domain.order._payment.entity;
+package poomasi.payment.entity;
 
 public enum PaymentMethod {
     KAKAO_PAY,
