+++ src/main/java/poomasi/payment/service/FarmPaymentService.java
@@ -1,0 +1,12 @@
+package poomasi.payment.service;
+
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class FarmPaymentService {
+
+
+
+
+}
